requirements.txt
cachelib==0.9.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Session==0.4.0
greenlet==1.1.3.post0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
requests==2.28.1
sqlparse==0.4.3
termcolor==2.0.1
urllib3==1.26.12
Werkzeug==2.2.2
